Simplified using statements for solution namespaces. Added player match rating images. Added PlayerDetails XML file type support. Added Last Match Rating column to PlayersView. Updated Database project.

+++ src/Hyperar.HPA.Database/dbo/Tables/League.sql
@@ -5,9 +5,9 @@
     [EnglishName]          NVARCHAR (256)  NOT NULL,
     [Continent]            NVARCHAR (256)  NOT NULL,
     [Zone]                 NVARCHAR (256)  NOT NULL,
-    [Season]               TINYINT         NOT NULL,
-    [Week]                 TINYINT         NOT NULL,
-    [SeasonOffset]         SMALLINT        NOT NULL,
+    [Season]               INT             NOT NULL,
+    [Week]                 INT             NOT NULL,
+    [SeasonOffset]         INT             NOT NULL,
     [LanguageId]           BIGINT          NOT NULL,
     [LanguageName]         NVARCHAR (256)  NOT NULL,
     [SeniorNationalTeamId] BIGINT          NOT NULL,
@@ -15,7 +15,7 @@
     [ActiveTeams]          INT             NOT NULL,
     [ActiveUsers]          INT             NOT NULL,
     [WaitingUsers]         INT             NOT NULL,
-    [NumberOfLevels]       TINYINT         NOT NULL,
+    [NumberOfLevels]       INT             NOT NULL,
     [NextTrainingUpdate]   DATETIME        NOT NULL,
     [NextEconomyUpdate]    DATETIME        NOT NULL,
     [NextCupMatchDate]     DATETIME        NULL,
@@ -23,6 +23,8 @@
     [FlagBytes]            VARBINARY (MAX) NOT NULL,
     CONSTRAINT [PK_League] PRIMARY KEY CLUSTERED ([HattrickId] ASC)
 );
+
+
