+++ src/Hyperar.HPA.Database/senior/Tables/HallOfFamePlayer.sql
@@ -1,20 +1,21 @@
 ﻿CREATE TABLE [senior].[HallOfFamePlayer] (
-    [HattrickId]               BIGINT         NOT NULL,
-    [FirstName]                NVARCHAR (256) NOT NULL,
-    [NickName]                 NVARCHAR (256) NULL,
-    [LastName]                 NVARCHAR (256) NOT NULL,
-    [Age]                      TINYINT         NOT NULL,
-    [JoinedTeamOn]             DATETIME       NOT NULL,
-    [NextBirthday]             DATETIME       NOT NULL,
-    [IntroducedOn] DATETIME       NOT NULL,
-    [ExpertType]               TINYINT        NOT NULL,
-    [CountryHattrickId]        BIGINT         NOT NULL,
-    [TeamHattrickId] BIGINT NOT NULL, 
+    [HattrickId]        BIGINT         NOT NULL,
+    [FirstName]         NVARCHAR (256) NOT NULL,
+    [NickName]          NVARCHAR (256) NULL,
+    [LastName]          NVARCHAR (256) NOT NULL,
+    [Age]               INT            NOT NULL,
+    [JoinedTeamOn]      DATETIME       NOT NULL,
+    [NextBirthday]      DATETIME       NOT NULL,
+    [IntroducedOn]      DATETIME       NOT NULL,
+    [ExpertType]        INT            NOT NULL,
+    [CountryHattrickId] BIGINT         NOT NULL,
+    [TeamHattrickId]    BIGINT         NOT NULL,
     CONSTRAINT [PK_HallOfFamePlayer] PRIMARY KEY CLUSTERED ([HattrickId] ASC),
     CONSTRAINT [FK_Senior_HallOfFamePlayer_Country] FOREIGN KEY ([CountryHattrickId]) REFERENCES [dbo].[Country] ([HattrickId]),
     CONSTRAINT [FK_Senior_HallOfFamePlayer_Team] FOREIGN KEY ([TeamHattrickId]) REFERENCES [senior].[Team] ([HattrickId])
-
 );
+
+
 
 
 GO
