+++ src/Hyperar.HPA.Database/senior/Tables/MatchTeamGoal.sql
@@ -1,16 +1,18 @@
 ﻿CREATE TABLE [senior].[MatchTeamGoal] (
     [Id]               INT            IDENTITY (1, 1) NOT NULL,
-    [Index]            TINYINT        NOT NULL,
+    [Index]            INT            NOT NULL,
     [PlayerHattrickId] BIGINT         NOT NULL,
     [PlayerName]       NVARCHAR (256) NOT NULL,
-    [HomeTeamScore]    TINYINT        NOT NULL,
-    [AwayTeamScore]    TINYINT        NOT NULL,
-    [Minute]           TINYINT        NOT NULL,
-    [MatchPart]        TINYINT        NOT NULL,
+    [HomeTeamScore]    INT            NOT NULL,
+    [AwayTeamScore]    INT            NOT NULL,
+    [Minute]           INT            NOT NULL,
+    [MatchPart]        INT            NOT NULL,
     [MatchTeamId]      INT            NOT NULL,
     CONSTRAINT [PK_MatchTeamGoal] PRIMARY KEY CLUSTERED ([Id] ASC),
     CONSTRAINT [FK_Senior_MatchTeamGoal_MatchTeam] FOREIGN KEY ([MatchTeamId]) REFERENCES [senior].[MatchTeam] ([Id])
 );
+
+
 
 
 GO
