+++ src/Hyperar.HPA.Database/senior/Tables/MatchTeamLineUpPlayer.sql
@@ -4,14 +4,16 @@
     [FirstName]         NVARCHAR (256) NOT NULL,
     [NickName]          NVARCHAR (256) NULL,
     [LastName]          NVARCHAR (256) NOT NULL,
-    [Role]              SMALLINT       NOT NULL,
-    [Behavior]          SMALLINT       NULL,
+    [Role]              INT            NOT NULL,
+    [Behavior]          INT            NULL,
     [Rating]            DECIMAL (3, 1) NULL,
     [EndRating]         DECIMAL (3, 1) NULL,
     [MatchTeamLineUpId] INT            NOT NULL,
     CONSTRAINT [PK_MatchTeamLineUpPlayer] PRIMARY KEY CLUSTERED ([Id] ASC),
     CONSTRAINT [FK_Senior_MatchTeamLineUpPlayer_MatchTeamLineUp] FOREIGN KEY ([MatchTeamLineUpId]) REFERENCES [senior].[MatchTeamLineUp] ([Id])
 );
+
+
 
 
 GO
