+++ src/Hyperar.HPA.Database/senior/Tables/MatchTeamLineUpStartingPlayer.sql
@@ -4,12 +4,14 @@
     [FirstName]         NVARCHAR (256) NOT NULL,
     [NickName]          NVARCHAR (256) NULL,
     [LastName]          NVARCHAR (256) NOT NULL,
-    [Role]              SMALLINT       NOT NULL,
-    [Behavior]          SMALLINT       NULL,
+    [Role]              INT            NOT NULL,
+    [Behavior]          INT            NULL,
     [MatchTeamLineUpId] INT            NOT NULL,
     CONSTRAINT [PK_MatchTeamLineUpStartingPlayer] PRIMARY KEY CLUSTERED ([Id] ASC),
     CONSTRAINT [FK_Senior_MatchTeamLineUpStartingPlayer_MatchTeamLineUp] FOREIGN KEY ([MatchTeamLineUpId]) REFERENCES [senior].[MatchTeamLineUp] ([Id])
 );
+
+
 GO
 
 CREATE NONCLUSTERED INDEX [IX_MatchTeamLineUpStartingPlayer_MatchTeamLineUpId]
