+++ src/Hyperar.HPA.Domain/Senior/StaffMember.cs
@@ -23,7 +23,7 @@
 
         public DateTime HiredOn { get; set; }
 
-        public byte Level { get; set; }
+        public int Level { get; set; }
 
         public string Name { get; set; }
