+++ src/Hyperar.HPA.Shared/Enums/MatchSectorRating.cs
@@ -1,6 +1,6 @@
 ﻿namespace Hyperar.HPA.Shared.Enums
 {
-    public enum MatchSectorRating : byte
+    public enum MatchSectorRating : int
     {
         DisastrousVeryLow = 1,
